refactor: inventorySearchService 제거 후 memberService에서 repository 추가
