+++ src/main/resources/static/docs/coupon.html
@@ -660,7 +660,7 @@ Content-Length: 66
 <div id="footer">
 <div id="footer-text">
 Version 0.0.1-SNAPSHOT<br>
-Last updated 2023-11-23 11:49:26 +0900
+Last updated 2023-11-23 15:00:05 +0900
 </div>
 </div>
 </body>
