+++ src/main/resources/static/docs/notification.html
@@ -515,7 +515,7 @@ Vary: Access-Control-Request-Headers</code></pre>
 <div id="footer">
 <div id="footer-text">
 Version 0.0.1-SNAPSHOT<br>
-Last updated 2023-11-23 11:49:26 +0900
+Last updated 2023-11-23 15:00:05 +0900
 </div>
 </div>
 </body>
